+++ src/led-strip-controller.h
@@ -9,7 +9,7 @@ enum ColorChannel { RED, GREEN, BLUE };
  * @param c - color channel enum (RED, GREEN, BLUE)
  * @param value - значение для канала в диапазоне от 0 до 1
  */
-void set_chanel_color(enum ColorChannel c, double value);
+void set_chanel_color(enum ColorChannel c, uint32_t value);
 
 /**
  * Функция устанавливающая цвет всей ленты
@@ -17,5 +17,5 @@ void set_chanel_color(enum ColorChannel c, double value);
  * @param g - значение для зелёного канала в диапазоне от 0 до 1
  * @param b - значение для синего канала в диапазоне от 0 до 1
  */
-void set_led_color(double red, double green, double blue);
+void set_led_color(uint32_t red,uint32_t green,uint32_t blue);
 #endif
